+++ frontend/src/pages/ManageUsers/ManageUsersPage.tsx
@@ -5,7 +5,6 @@ export default function ManageUsersPage() {
   useTopNavBarTitle("Manage Users");
   return (
     <div className="container-fluid my-4">
-      <h2>Manage Users</h2>
       <TableManageUsers />
     </div>
   );